black formatting, flake linting, mypy corrections across files. Added test for tf_example_serialization, added / corrected tests for other data / train files.

+++ basic_pitch/data/tf_example_serialization.py
@@ -49,9 +49,9 @@ def _to_transcription_tfex(
     encoded_wav: bytes,
     notes_indices: List[Tuple[int, int]],
     notes_values: List[float],
-    onsets_indices: List[float],
+    onsets_indices: List[int],
     onsets_values: List[float],
-    contours_indices: List[float],
+    contours_indices: List[int],
     contours_values: List[float],
     notes_onsets_shape: Tuple[int, int],
     contours_shape: Tuple[int, int],
@@ -62,30 +62,14 @@ def _to_transcription_tfex(
                 "file_id": bytes_feature(bytes(file_id, "utf-8")),
                 "source": bytes_feature(bytes(source, "utf-8")),
                 "audio_wav": bytes_feature(encoded_wav),
-                "notes_indices": bytes_feature(
-                    tf.io.serialize_tensor(np.array(notes_indices, np.int64))
-                ),
-                "notes_values": bytes_feature(
-                    tf.io.serialize_tensor(np.array(notes_values, np.float32))
-                ),
-                "onsets_indices": bytes_feature(
-                    tf.io.serialize_tensor(np.array(onsets_indices, np.int64))
-                ),
-                "onsets_values": bytes_feature(
-                    tf.io.serialize_tensor(np.array(onsets_values, np.float32))
-                ),
-                "contours_indices": bytes_feature(
-                    tf.io.serialize_tensor(np.array(contours_indices, np.int64))
-                ),
-                "contours_values": bytes_feature(
-                    tf.io.serialize_tensor(np.array(contours_values, np.float32))
-                ),
-                "notes_onsets_shape": bytes_feature(
-                    tf.io.serialize_tensor(np.array(notes_onsets_shape, np.int64))
-                ),
-                "contours_shape": bytes_feature(
-                    tf.io.serialize_tensor(np.array(contours_shape, np.int64))
-                ),
+                "notes_indices": bytes_feature(tf.io.serialize_tensor(np.array(notes_indices, np.int64))),
+                "notes_values": bytes_feature(tf.io.serialize_tensor(np.array(notes_values, np.float32))),
+                "onsets_indices": bytes_feature(tf.io.serialize_tensor(np.array(onsets_indices, np.int64))),
+                "onsets_values": bytes_feature(tf.io.serialize_tensor(np.array(onsets_values, np.float32))),
+                "contours_indices": bytes_feature(tf.io.serialize_tensor(np.array(contours_indices, np.int64))),
+                "contours_values": bytes_feature(tf.io.serialize_tensor(np.array(contours_values, np.float32))),
+                "notes_onsets_shape": bytes_feature(tf.io.serialize_tensor(np.array(notes_onsets_shape, np.int64))),
+                "contours_shape": bytes_feature(tf.io.serialize_tensor(np.array(contours_shape, np.int64))),
             }
         )
     )
@@ -97,9 +81,9 @@ def to_transcription_tfexample(
     audio_wav_file_path: str,
     notes_indices: List[Tuple[int, int]],
     notes_values: List[float],
-    onsets_indices: List[float],
+    onsets_indices: List[int],
     onsets_values: List[float],
-    contours_indices: List[float],
+    contours_indices: List[int],
     contours_values: List[float],
     notes_onsets_shape: Tuple[int, int],
     contours_shape: Tuple[int, int],
